Tests and delete users from org flow
